Refactor & add granularity filter
