+++ dashboard/client/filter/date-range/date-range.html
@@ -1,0 +1,10 @@
+<template name="filterDateRange">
+  <div class="form-group-sm">
+    <div class="input-group">
+      <div class="input-group-addon">From</div>
+      <input type="date" class="form-control" id="analytics-date-from" >
+      <div class="input-group-addon">To</div>
+      <input type="date" class="form-control" id="analytics-date-to" >
+    </div>
+  </div>
+</template>
