+++ dashboard/client/filter/granularity/granularity.html
@@ -1,0 +1,23 @@
+<template name="filterGranularity">
+  <div class="btn-group btn-group-sm" data-toggle="buttons" id="date-granularity-selector">
+    <label class="btn btn-default active" for="month-input" id="month-granularity">
+      <input type="radio" name="granularity" value="month">
+      Month
+    </label>
+
+    <label class="btn btn-default" for="week-input" id="week-granularity">
+      <input type="radio" name="granularity" value="week">
+      Week
+    </label>
+
+    <label class="btn btn-default" for="day-input" id="day-granularity">
+      <input type="radio" name="granularity" value="day">
+      Day
+    </label>
+
+    <label class="btn btn-default" for="hour-input" id="hour-granularity">
+      <input type="radio" name="granularity" value="hour">
+      Hour
+    </label>
+  </div>
+</template>
