+++ package.json
@@ -8,7 +8,7 @@
     "babel-runtime": "^6.23.0",
     "d3": "^3.5.17",
     "elasticsearch": "^12.0.1",
-    "jquery": "^3.1.1",
+    "jquery": "^2.2.4",
     "lodash": "^4.17.2",
     "meteor-node-stubs": "~0.2.0",
     "moment": "^2.16.0",
